# api/openapi.yaml
openapi: 3.1.0
info:
  title: Omega control plane API
  version: 0.0.1
  summary: SPIFFE-compatible workload identity + OpenID AuthZEN 1.0 PDP.
  description: |
    HTTP/JSON surface served by `omega server` on `--http-addr` (default
    `127.0.0.1:8080`). The gRPC SPIFFE Workload API is served by
    `omega agent` on a Unix socket and is not described here.

    ## Leader-only endpoints

    When Omega runs against Postgres with HA leader election, every
    write endpoint plus `POST /access/v1/evaluation` is gated by the
    advisory-lock leader state. A non-leader replica responds with
    `503 Service Unavailable` and `Retry-After: 1`. The endpoints
    affected are tagged `leader-only` below.

    ## Audit log

    Every successful write endpoint and every authorization decision
    appends an entry to the tamper-evident audit log
    (`GET /v1/audit`). Hash chain integrity can be verified with
    `GET /v1/audit/verify`.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
  contact:
    name: Omega maintainers
    url: https://github.com/kanywst/omega
servers:
  - url: http://127.0.0.1:8080
    description: Default development listen address (single-binary quickstart).
...
paths:
...
  /v1/domains:
    get:
      tags: [domains]
      operationId: listDomains
      summary: List domains.
      responses:
        "200":
          description: Domain list (possibly empty).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DomainListResponse"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [domains, leader-only]
      operationId: createDomain
      summary: Create a SPIFFE namespace.
      description: |
        The name is dot-separated lowercase labels (`media.news`), and
        its parent (`media`) must already exist. Under `--require-auth`
        the caller must administer the parent, or be a
        `--domain-root-admin` for a top-level domain. With no `admins`
        given, the authenticated caller becomes the domain's admin.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DomainCreateRequest"
      responses:
        "201":
          description: Domain created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Domain"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "503":
          $ref: "#/components/responses/NotLeader"
  /v1/domains/{name}:
    parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
        description: Domain name (e.g. `example`).
    get:
      tags: [domains]
      operationId: getDomain
      summary: Fetch a single domain.
      responses:
        "200":
          description: Domain.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Domain"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags: [domains, leader-only]
      operationId: deleteDomain
      summary: Delete a domain that has no child domains.
      description: |
        Under `--require-auth` the caller must administer the domain's
        parent (or be a root admin for a top-level domain); a domain's
        own admins cannot delete it.
      responses:
        "204":
          description: Deleted, with its admin grants.
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "503":
          $ref: "#/components/responses/NotLeader"
  /v1/domains/{name}/admins:
    parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
    post:
      tags: [domains, leader-only]
      operationId: addDomainAdmin
      summary: Grant a SPIFFE ID admin rights on a domain and every domain below it.
      description: Under `--require-auth` the caller must administer the domain or an ancestor.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DomainAdminRequest"
      responses:
        "200":
          description: The domain with its updated admins.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Domain"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/NotLeader"
    delete:
      tags: [domains, leader-only]
      operationId: removeDomainAdmin
      summary: Revoke a SPIFFE ID's admin grant on a domain.
      description: |
        Revokes the grant on this domain only. Grants the principal
        holds on domains below it are separate and stay.
      parameters:
        - name: principal
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The domain with its updated admins.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Domain"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/NotLeader"
  /v1/domains/{name}/groups:
    parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [domains]
      operationId: listGroups
      summary: List a domain's groups.
      description: Member lists are omitted for unauthenticated callers under `--require-auth`.
      responses:
        "200":
          description: Group list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
    post:
      tags: [domains, leader-only]
      operationId: createGroup
      summary: Create a group owned by the domain.
      description: Under `--require-auth` the caller must administer the domain or an ancestor.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupRequest"
      responses:
        "201":
          description: Group created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "503":
          $ref: "#/components/responses/NotLeader"
...
components:
  schemas:
...
    Error:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message.
...
    Domain:
      type: object
      additionalProperties: false
      required: [name, created_at]
      properties:
        name:
          type: string
          examples: [example]
        parent:
          type: string
          description: Parent domain name (omitted for top-level domains).
        description:
          type: string
        admins:
          type: array
          items:
            type: string
          description: SPIFFE IDs that administer this domain and every domain below it.
        created_at:
          type: string
          format: date-time
    DomainCreateRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name:
          type: string
        parent:
          type: string
          description: Optional; must equal the parent implied by the name.
        description:
          type: string
        admins:
          type: array
          items:
            type: string
    DomainAdminRequest:
      type: object
      additionalProperties: false
      required: [principal]
      properties:
        principal:
          type: string
          description: SPIFFE ID to grant.
    Group:
      type: object
      additionalProperties: false
      required: [domain, name, created_at]
      properties:
        domain:
          type: string
        name:
          type: string
          description: Policy identity is Group::"<domain>:<name>".
        description:
          type: string
        members:
          type: array
          items:
            $ref: "#/components/schemas/GroupMember"
        created_at:
          type: string
          format: date-time
    GroupMember:
      type: object
      additionalProperties: false
      required: [principal]
      properties:
        principal:
          type: string
        expires_at:
          type: string
          format: date-time
          description: Omitted for a membership that never expires. Checked at evaluation time.
    GroupRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
...
    GroupListResponse:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Group"
    DomainListResponse:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Domain"
...
  responses:
    BadRequest:
      description: Validation failure.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Authorization failure.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource does not exist.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Resource already exists.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Server-side error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotLeader:
      description: |
        This replica is not the Postgres advisory-lock leader. Retry
        against the leader (`GET /v1/leader`) or, on transient
        failover, after the `Retry-After` window.
      headers:
        Retry-After:
          schema:
            type: string
            const: "1"
          description: Seconds to wait before retrying.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"